+++ lib/array/any/method.ts
@@ -1,0 +1,3 @@
+const method = (arr: any[], fn: (value: any, index: number, array: any[]) => boolean = Boolean) => arr.some(fn);
+
+export = method;