+++ lib/array/clone/index.ts
@@ -1,8 +1,10 @@
-export { }
+import * as method from "./method";
+
+export { };
 
 declare global {
   interface Array<T> {
-    clone(): Array<T>
+    clone(): T[];
   }
 }
 
@@ -13,6 +15,6 @@ declare global {
  * @example
  * [1, 2, 3].clone(); // [1, 2, 3]
  */
-Array.prototype.clone = function(): Array<any> {
-  return [...this]
-}
+Array.prototype.clone = function() {
+  return method(this);
+};