+++ lib/object/invert/index.ts
@@ -4,7 +4,7 @@ export { }
 
 declare global {
   interface Object {
-    invert(): object
+    $invert(): object
   }
 }
 
@@ -13,9 +13,9 @@ declare global {
  * @memberof Object.prototype
  * @returns {Object}
  * @example
- * { name: 'John', age: 20 }.invert(); // { 20: 'age', John: 'name' }
+ * { name: 'John', age: 20 }.$invert(); // { 20: 'age', John: 'name' }
  */
-function invert(this: { [key: string]: any }): object {
+function $invert(this: { [key: string]: any }): object {
   return Object.keys(this).reduce((acc: { [key: string]: any }, key) => {
     acc[this[key]] = key
 
@@ -23,4 +23,4 @@ function invert(this: { [key: string]: any }): object {
   }, {})
 }
 
-addPrototype('invert', invert)
+addPrototype('$invert', $invert)