+++ lib/string/index.ts
@@ -1,5 +1,7 @@
 /** @namespace String */
 
+import './base64'
+import './base64Decode'
 import './camelCase'
 import './capitalize'
 import './chars'
@@ -9,6 +11,7 @@ import './humanize'
 import './isInstance'
 import './kebabCase'
 import './lines'
+import './map'
 import './mask'
 import './pluralize'
 import './reverse'