+++ lib/string/map/method.ts
@@ -1,0 +1,6 @@
+const map = (
+  str: string,
+  fn: (char: string, index: number, chars: string[]) => string,
+): string => str.split("").map(fn).join("");
+
+export = map;